Move pull request into repository module
